Restricted some pages (auth)

+++ products/views.py
@@ -1,5 +1,5 @@
-from django.shortcuts import render, get_object_or_404
-
+from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
 from .models import Product
 from .forms import AddProductForm
 
@@ -18,30 +18,41 @@ def products_details(request, pk):
     return render(request, 'products/product-details.html', params)
 
 
+@login_required
 def product_add(request):
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save()
-            return render(request, "products/product-added.html")
+    if request.user.is_authenticated and request.user.is_superuser:
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()
+                return render(request, "products/product-added.html")
+        else:
+            form = AddProductForm()
+        return render(request, "products/product-add.html", {'form': form})
     else:
-        form = AddProductForm()
-    return render(request, "products/product-add.html", {'form': form})
+        return redirect('products_list')
 
 
 def product_edit(request, pk):
-    product = get_object_or_404(Product, pk=pk)
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
 
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES, instance=product)
-        if form.is_valid():
-            form.save()
-            return render(request, "products/product-added.html")
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES, instance=product)
+            if form.is_valid():
+                form.save()
+                return render(request, "products/product-added.html")
+        else:
+            form = AddProductForm(instance=product)
+        return render(request, "products/product-add.html", {'form': form})
     else:
-        form = AddProductForm(instance=product)
-    return render(request, "products/product-add.html", {'form': form})
+        return redirect('products_list')
 
 
 def product_delete(request, pk):
-    get_object_or_404(Product, pk=pk).delete()
-    return render(request, "products/product-deleted.html")
+    if request.user.is_authenticated and request.user.is_superuser:
+        get_object_or_404(Product, pk=pk).delete()
+        return render(request, "products/product-deleted.html")
+    else:
+        return redirect('products_list')
+
